shared method to lookupStrongsNumbers()
